Enum biscuit::errors::Error

pub enum Error {
    GenericError(String),
    ValidationError(ValidationError),
    JsonError(Error),
    DecodeBase64(DecodeError),
    Utf8(Utf8Error),
    IOError(Error),
    UriParseError(ParseError),
    WrongKeyType {
        expected: String,
        actual: String,
    },
    WrongEncryptionOptions {
        expected: String,
        actual: String,
    },
    UnspecifiedCryptographicError,
    UnsupportedOperation,
}

All the errors we can encounter while signing/verifying tokens and a couple of custom one for when the token we are trying to verify is invalid

Variants
GenericError(String)

A generic error which is described by the contained string

ValidationError(ValidationError)

Error returned from failed token validation

JsonError(Error)

Error during the serialization or deserialization of tokens

DecodeBase64(DecodeError)

Error during base64 encoding or decoding

Utf8(Utf8Error)

Error when decoding bytes to UTF8 string

IOError(Error)

Errors related to IO

UriParseError(ParseError)

Errors related to URI parsing

WrongKeyType

Wrong key type was provided for the cryptographic operation

Fields of WrongKeyType

expected: String

Expected type of key

actual: String

Actual type of key

WrongEncryptionOptions

Wrong variant of EncryptionOptions was provided for the encryption operation

Fields of WrongEncryptionOptions

expected: String

Expected variant of options

actual: String

Actual variant of options

UnspecifiedCryptographicError

An unknown cryptographic error

UnsupportedOperation

An unsupported or invalid operation