Struct biscuit::jwe::Header

pub struct Header<T> {
    pub registered: RegisteredHeader,
    pub cek_algorithm: CekAlgorithmHeader,
    pub private: T,
}

JWE Header, consisting of the registered fields and other custom fields

Fields
registered: RegisteredHeader

Registered header fields

cek_algorithm: CekAlgorithmHeader

Key management algorithm specific headers

private: T

Private header fields

Methods

impl Header<Empty>

pub fn from_registered_header(registered: RegisteredHeader) -> Self

Convenience function to create a header with only registered headers