Enum biscuit::jwk::KeyOperations

pub enum KeyOperations {
    Sign,
    Verify,
    Encrypt,
    Decrypt,
    WrapKey,
    UnwrapKey,
    DeriveKey,
    DeriveBits,
    Other(String),
}

Operations that the key is intended to be used for. This enum is serialized untagged

Variants
Sign

Computer digitial signature or MAC

Verify

Verify digital signature or MAC

Encrypt

Encrypt content

Decrypt

Decrypt content and validate decryption, if applicable

WrapKey

Encrypt key

UnwrapKey

Decrypt key and validate decryption, if applicable

DeriveKey

Derive key

DeriveBits

Derive bits not to be used as a key

Other(String)

Other operation