Struct biscuit::jwk::CommonParameters

pub struct CommonParameters {
    pub public_key_use: Option<PublicKeyUse>,
    pub key_operations: Option<Vec<KeyOperations>>,
    pub algorithm: Option<Algorithm>,
    pub key_id: Option<String>,
    pub x509_url: Option<String>,
    pub x509_chain: Option<Vec<String>>,
    pub x509_fingerprint: Option<String>,
}

Common JWK parameters

Fields
public_key_use: Option<PublicKeyUse>

The intended use of the public key. Should not be specified with key_operations. See sections 4.2 and 4.3 of RFC7517.

key_operations: Option<Vec<KeyOperations>>

The "key_ops" (key operations) parameter identifies the operation(s) for which the key is intended to be used. The "key_ops" parameter is intended for use cases in which public, private, or symmetric keys may be present. Should not be specified with public_key_use. See sections 4.2 and 4.3 of RFC7517.

algorithm: Option<Algorithm>

The algorithm intended for use with the key

key_id: Option<String>

The case sensitive Key ID for the key

x509_url: Option<String>

X.509 Public key cerfificate URL. This is currently not implemented (correctly). Serialized to x5u.

x509_chain: Option<Vec<String>>

X.509 public key certificate chain. This is currently not implemented (correctly). Serialized to x5c.

x509_fingerprint: Option<String>

X.509 Certificate thumbprint. This is currently not implemented (correctly). Also not implemented, is the SHA-256 thumbprint variant of this header. Serialized to x5t.