Struct biscuit::jwk::JWK

pub struct JWK<T> {
    pub common: CommonParameters,
    pub algorithm: AlgorithmParameters,
    pub additional: T,
}

A JSON object that represents a cryptographic key. The members of the object represent properties of the key, including its value. Type T is a struct representing additional JWK properties

Fields
common: CommonParameters

Common JWK parameters

algorithm: AlgorithmParameters

Key algorithm specific parameters

additional: T

Additional JWK parameters

Methods

impl<T: Serialize + DeserializeOwned> JWK<T>

pub fn new_octect_key(key: &[u8], additional: T) -> Self

Convenience to create a new bare-bones Octect key

pub fn clone_without_additional(&self) -> JWK<Empty>

Convenience function to strip out the additional fields

pub fn key_type(&self) -> KeyType

Returns the type of key represented by this key

pub fn octect_key(&self) -> Result<&[u8], Error>

Return the byte sequence of an octect key