Struct biscuit::jwk::RSAKeyParameters

pub struct RSAKeyParameters {
    pub key_type: RSAKeyType,
    pub n: BigUint,
    pub e: BigUint,
    pub d: Option<BigUint>,
    pub p: Option<BigUint>,
    pub q: Option<BigUint>,
    pub dp: Option<BigUint>,
    pub dq: Option<BigUint>,
    pub qi: Option<BigUint>,
    pub other_primes_info: Option<Vec<OtherPrimesInfo>>,
}

Parameters for a RSA Key

Fields
key_type: RSAKeyType

Key type value for a RSA Key

n: BigUint

The "n" (modulus) parameter contains the modulus value for the RSA public key. It is serialized as a Base64urlUInt-encoded value.

e: BigUint

The "e" (exponent) parameter contains the exponent value for the RSA public key. It is serialized as a Base64urlUInt-encoded value.

d: Option<BigUint>

The "d" (private exponent) parameter contains the private exponent value for the RSA private key. It is serialized as a Base64urlUInt-encoded value.

p: Option<BigUint>

The "p" (first prime factor) parameter contains the first prime factor. It is serialized as a Base64urlUInt-encoded value.

q: Option<BigUint>

The "q" (second prime factor) parameter contains the second prime factor. It is serialized as a Base64urlUInt-encoded value.

dp: Option<BigUint>

The "dp" (first factor CRT exponent) parameter contains the Chinese Remainder Theorem (CRT) exponent of the first factor. It is serialized as a Base64urlUInt-encoded value.

dq: Option<BigUint>

The "dq" (second factor CRT exponent) parameter contains the CRT exponent of the second factor. It is serialized as a Base64urlUInt-encoded value.

qi: Option<BigUint>

The "qi" (first CRT coefficient) parameter contains the CRT coefficient of the second factor It is serialized as a Base64urlUInt-encoded value.

other_primes_info: Option<Vec<OtherPrimesInfo>>

The "oth" (other primes info) parameter contains an array of information about any third and subsequent primes, should they exist.