Module biscuit::jws

JSON Web Signatures, including JWT signing and headers

Defined in RFC 7515. For most common use, you will want to look at the Compact enum.

Structs

Header
JWS Header, consisting of the registered fields and other custom fields

RegisteredHeader
Registered JWS header fields. The alg defaults to HS256 and typ is automatically set to JWT. All the other fields are optional. The fields are defined by RFC7519#5 and additionally in RFC7515#4.1.

Enums

Compact
Compact representation of a JWS

Secret
The secrets used to sign and/or encrypt tokens