Enum biscuit::jws::Compact

pub enum Compact<T, H> {
    Decoded {
        header: Header<H>,
        payload: T,
    },
    Encoded(Compact),
}

Compact representation of a JWS

This representation contains a payload (type T) (e.g. a claims set) and is (optionally) signed. This is the most common form of tokens used. The JWS can contain additional header fields provided by type H.

Serialization/deserialization is handled by serde. Before you transport the token, make sure you turn it into the encoded form first.

Examples

Signing and verifying a JWT with HS256

See an example in the biscuit::JWT type alias.

Variants
Decoded

Decoded form of the JWS. This variant cannot be serialized or deserialized and will return an error.

Fields of Decoded

header: Header<H>

Embedded header

payload: T

Payload, usually a claims set

Encoded(Compact)

Encoded and (optionally) signed JWT. Use this form to send to your clients

Methods

impl<T, H> Compact<T, H> where
T: CompactPart,
H: Serialize + DeserializeOwned,

pub fn new_decoded(header: Header<H>, payload: T) -> Self

New decoded JWT

pub fn new_encoded(token: &str) -> Self

New encoded JWT

pub fn into_encoded(self, secret: &Secret) -> Result<Self, Error>

Consumes self and convert into encoded form. If the token is already encoded, this is a no-op.

pub fn encode(&self, secret: &Secret) -> Result<Self, Error>

Encode the JWT passed and sign the payload using the algorithm from the header and the secret The secret is dependent on the signing algorithm

pub fn into_decoded(
self,
secret: &Secret,
algorithm: SignatureAlgorithm
) -> Result<Self, Error>

Consumes self and convert into decoded form, verifying the signature, if any. If the token is already decoded, this is a no-op

pub fn decode(
&self,
secret: &Secret,
algorithm: SignatureAlgorithm
) -> Result<Self, Error>

Decode a token into the JWT struct and verify its signature If the token or its signature is invalid, it will return an error

pub fn encoded(&self) -> Result<&Compact, Error>

Convenience method to get a reference to the encoded string from an encoded compact JWS

pub fn encoded_mut(&mut self) -> Result<&mut Compact, Error>

Convenience method to get a mutable reference to the encoded string from an encoded compact JWS

pub fn payload(&self) -> Result<&T, Error>

Convenience method to get a reference to the claims set from a decoded compact JWS

pub fn payload_mut(&mut self) -> Result<&mut T, Error>

Convenience method to get a reference to the claims set from a decoded compact JWS

pub fn header(&self) -> Result<&Header<H>, Error>

Convenience method to get a reference to the header from a decoded compact JWS

pub fn header_mut(&mut self) -> Result<&mut Header<H>, Error>

Convenience method to get a reference to the header from a decoded compact JWS

pub fn unwrap_decoded(self) -> (Header<H>, T)

Consumes self, and move the payload and header out and return them as a tuple

Panics

Panics if the JWS is not decoded

pub fn unwrap_encoded(self) -> Compact

Consumes self, and move the encoded Compact out and return it

Panics

Panics if the JWS is not encoded

pub fn unverified_header(&self) -> Result<Header<H>, Error>

Without decoding and verifying the JWS, retrieve a copy of the header.

Warning

Use this at your own risk. It is not advisable to trust unverified content.

pub fn unverified_payload(&self) -> Result<T, Error>

Without decoding and verifying the JWS, retrieve a copy of the payload.

Warning

Use this at your own risk. It is not advisable to trust unverified content.

pub fn signature(&self) -> Result<Vec<u8>, Error>

Get a copy of the signature

impl<P, H> Compact<ClaimsSet<P>, H> where
ClaimsSet<P>: CompactPart,
H: Serialize + DeserializeOwned,

Convenience implementation for a Compact that contains a ClaimsSet

pub fn validate_times(
&self,
options: Option<TemporalValidationOptions>
) -> Result<(), Error>

Validate the temporal claims in the decoded token

If None is provided for options, the defaults will apply.

By default, no temporal claims (namely iat, exp, nbf) are required, and they will pass validation if they are missing.