Struct biscuit::Compact

pub struct Compact {
    pub parts: Vec<Base64Url>,
}

A collection of CompactParts that have been converted to Base64Url

Fields
parts: Vec<Base64Url>

Parts of the compact representation

Methods

impl Compact

pub fn new() -> Self

Create an empty struct

pub fn with_capacity(capacity: usize) -> Self

Create an empty struct with some expected capacity

pub fn push(&mut self, part: &CompactPart) -> Result<(), Error>

Push a CompactPart to the end

pub fn len(&self) -> usize

Returns the number of parts

pub fn is_empty(&self) -> bool

Returns whether there are no parts

pub fn encode(&self) -> String

Encodes the various parts into Base64 URL encoding and then concatenates them with period '.' This corresponds to the various Compact representation in JWE and JWS, for example

pub fn decode(encoded: &str) -> Self

Convenience function to split an encoded compact representation into a list of Base64Url.

pub fn part<T: CompactPart>(&self, index: usize) -> Result<T, Error>

Convenience function to retrieve a part at a certain index and decode into the type desired

pub fn to_string(&self) -> String

Alias of encode