Trait biscuit::CompactJson

pub trait CompactJson: Serialize + DeserializeOwned { }

A marker trait that indicates that the object is to be serialized to JSON and deserialized from JSON. This is primarily used in conjunction with the CompactPart trait which will serialize structs to JSON before base64 encoding, and vice-versa.