Enum biscuit::SingleOrMultiple

pub enum SingleOrMultiple<T> {
    Single(T),
    Multiple(Vec<T>),
}

Represents a choice between a single value or multiple values. This value is serialized by serde untagged.

Examples

use biscuit::SingleOrMultiple;
use serde::{Serialize, Deserialize};

#[derive(Debug, Eq, PartialEq, Serialize, Deserialize)]
struct SingleOrMultipleStrings {
    values: SingleOrMultiple<String>,
}

let single = SingleOrMultipleStrings {
    values: SingleOrMultiple::Single("foobar".to_string())
};
let expected_json = r#"{"values":"foobar"}"#;

let serialized = serde_json::to_string(&single).unwrap();
assert_eq!(expected_json, serialized);

let deserialized: SingleOrMultipleStrings = serde_json::from_str(&serialized).unwrap();
assert_eq!(deserialized, single);

let multiple = SingleOrMultipleStrings {
    values: SingleOrMultiple::Multiple(vec!["foo".to_string(),
                                            "bar".to_string(),
                                            "baz".to_string()]),
};

let expected_json = r#"{"values":["foo","bar","baz"]}"#;

let serialized = serde_json::to_string(&multiple).unwrap();
assert_eq!(expected_json, serialized);

let deserialized: SingleOrMultipleStrings = serde_json::from_str(&serialized).unwrap();
assert_eq!(deserialized, multiple);

Variants

Single(T)

One single value

Multiple(Vec<T>)

Multiple values

Implementations

impl<T> SingleOrMultiple<T> where
T: Clone + Debug + Eq + PartialEq + Serialize + DeserializeOwned + Send + Sync,

pub fn contains<Q: ?Sized>(&self, value: &Q) -> bool where
T: Borrow<Q>,
Q: PartialEq,

Checks whether this enum, regardless of single or multiple value contains value.

pub fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = &'a T> + 'a>

Yields an iterator for the single value or the list