Module biscuit::jwa

JSON Web Algorithms

Code for implementing JWA according to RFC 7518.

Typically, you will not use these directly, but as part of a JWS or JWE.

Structs

EncryptionResult
The result returned from an encryption operation

Enums

Algorithm
Algorithms described by RFC 7518. This enum is serialized untagged.

ContentEncryptionAlgorithm
Algorithms meant for content encryption. See RFC7518#5

EncryptionOptions
Options to be passed in while performing an encryption operation, if required by the algorithm.

KeyManagementAlgorithm
Algorithms for key management as defined in RFC7518#4

KeyManagementAlgorithmType
Describes the type of operations that the key management algorithm supports with respect to a Content Encryption Key (CEK)

SignatureAlgorithm
The algorithms supported for digital signature and MACs, defined by RFC7518#3.

Traits

SecureRandom
A secure random number generator.