Struct biscuit::jwe::RegisteredHeader

pub struct RegisteredHeader {
    pub cek_algorithm: KeyManagementAlgorithm,
    pub enc_algorithm: ContentEncryptionAlgorithm,
    pub compression_algorithm: Option<CompressionAlgorithm>,
    pub media_type: Option<String>,
    pub content_type: Option<String>,
    pub web_key_url: Option<String>,
    pub web_key: Option<String>,
    pub key_id: Option<String>,
    pub x509_url: Option<String>,
    pub x509_chain: Option<Vec<String>>,
    pub x509_fingerprint: Option<String>,
    pub critical: Option<Vec<String>>,
}

Registered JWE header fields. The fields are defined by RFC 7516#4.1

Fields
cek_algorithm: KeyManagementAlgorithm

Algorithm used to encrypt or determine the value of the Content Encryption Key

enc_algorithm: ContentEncryptionAlgorithm

Content encryption algorithm used to perform authenticated encryption on the plaintext to produce the ciphertext and the Authentication Tag

compression_algorithm: Option<CompressionAlgorithm>

Compression algorithm applied to plaintext before encryption, if any. Compression is not supported at the moment. Must only appear in integrity protected header.

media_type: Option<String>

Media type of the complete JWE. Serialized to typ. Defined in RFC7519#5.1 and additionally RFC7515#4.1.9. The "typ" value "JOSE" can be used by applications to indicate that this object is a JWS or JWE using the JWS Compact Serialization or the JWE Compact Serialization. The "typ" value "JOSE+JSON" can be used by applications to indicate that this object is a JWS or JWE using the JWS JSON Serialization or the JWE JSON Serialization. Other type values can also be used by applications.

content_type: Option<String>

Content Type of the secured payload. Typically used to indicate the presence of a nested JOSE object which is signed or encrypted. Serialized to cty. Defined in RFC7519#5.2 and additionally RFC7515#4.1.10.

web_key_url: Option<String>

The JSON Web Key Set URL. This is currently not implemented (correctly). Serialized to jku. Defined in RFC7515#4.1.2.

web_key: Option<String>

The JSON Web Key. This is currently not implemented (correctly). Serialized to jwk. Defined in RFC7515#4.1.3.

key_id: Option<String>

The Key ID. This is currently not implemented (correctly). Serialized to kid. Defined in RFC7515#4.1.3.

x509_url: Option<String>

X.509 Public key cerfificate URL. This is currently not implemented (correctly). Serialized to x5u. Defined in RFC7515#4.1.5.

x509_chain: Option<Vec<String>>

X.509 public key certificate chain. This is currently not implemented (correctly). Serialized to x5c. Defined in RFC7515#4.1.6.

x509_fingerprint: Option<String>

X.509 Certificate thumbprint. This is currently not implemented (correctly). Also not implemented, is the SHA-256 thumbprint variant of this header. Serialized to x5t. Defined in RFC7515#4.1.7.

critical: Option<Vec<String>>

List of critical extended headers. This is currently not implemented (correctly). Serialized to crit. Defined in RFC7515#4.1.11.