Struct biscuit::jwk::EllipticCurveKeyParameters

pub struct EllipticCurveKeyParameters {
    pub key_type: EllipticCurveKeyType,
    pub curve: EllipticCurve,
    pub x: Vec<u8>,
    pub y: Vec<u8>,
    pub d: Option<Vec<u8>>,
}

Parameters for an Elliptic Curve Key

Fields
key_type: EllipticCurveKeyType

Key type value for an Elliptic Curve Key.

curve: EllipticCurve

The "crv" (curve) parameter identifies the cryptographic curve used with the key.

x: Vec<u8>

The "x" (x coordinate) parameter contains the x coordinate for the Elliptic Curve point. Serialized to base64 URL encoded

y: Vec<u8>

The "y" (y coordinate) parameter contains the y coordinate for the Elliptic Curve point. Serialized to base64 URL encoded

d: Option<Vec<u8>>

The "d" (ECC private key) parameter contains the Elliptic Curve private key value.